♻️ Rename the module as abstraction to be more clear.
