+++ gearmeshing_ai/info_provider/prompt/__init__.py
@@ -1,6 +1,6 @@
 from .builtin import BuiltinPromptProvider
 from .loader import load_prompt_provider
-from .provider import PromptProvider
+from .base import PromptProvider
 from .reload import HotReloadWrapper
 from .stacked import StackedPromptProvider
