+++ gearmeshing_ai/info_provider/prompt/stacked.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Optional
 
-from .provider import PromptProvider
+from .base import PromptProvider
 
 
 class StackedPromptProvider(PromptProvider):
